define new approach that does not use exhaustive run*
Generate new rules until we have a fully grounded solution. This does not
exhaustively generate possible solutions from run+ making individual iterations
faster; however, that means we cannot avoid adding rules that do not reduce
the solution space.

Still not improve it much. The best way to generate the rule set would be to
have a specific real expected solution and derive rules against that. Originally,
I didn't want to give up the flexibility of not being able to change things as I
went. This is because I actually expect to use this in an iterative way.